pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["Temirlan Ilik <0penguin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-money = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
